fix timout by using our own input buffer
The timeout was improperly implemented using select() on a buffered file
handle. This cannot work because there will be data to use in the buffer even
if the handle is not ready to read, thus we deadlock.

We must use our own buffering so we know to call select() only when we are
certain we need to fetch more data from the handle.