Handle non-blocking write
The I/O handle is marked as non-blocking, so syswrite will on occasion return
EAGAIN. This is now handled. I also handle EINTR since that's always possible
and was cheap to test.

This code as committed is correct, however I'm not 100% satisified with the
structure of having some I/O in NATS::Client and some in NATS::Connection.